Enable progress reporting for ios.
Modifies the SimpleSessionExporter to expose a progress Float value. This Float value defaults to 0 if there is no exportSession. A new exportSession private var was added to store the current exportSession.
The AVAssetExportSession completion handler updates the exportSession reference to nil once it has completed.

To actually send the updates, we add a schedule job every 250 milliseconds which calls our progressHandler. The exporter completion handler cancels it when it is done.